pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.0",
    "nltk>=3.9.1",
    "pillow>=11.2.1",
    "pypdf2>=3.0.1",
    "pytesseract>=0.3.13",
    "scikit-learn>=1.6.1",
    "werkzeug>=3.1.3",
]
